+++ Route.h
@@ -9,7 +9,7 @@ class Route {
     const Airport& destination_;
     std::string airline_code;
     int stops_;
-    int weight_;
+    double weight_;
 
    public:
     Route(const Airport& source, const Airport& destination,
@@ -20,6 +20,6 @@ class Route {
     const Airport& get_destination();
     std::string get_code();
     int get_stops();
-    int get_dist();
-    int get_weight();
+    double get_dist();
+    double get_weight();
 };
